moved net consts into common grouped all common net code inside pkNet namespace
