--- server/netconsts.h
@@ -1,7 +1,0 @@
-#pragma once
-
-namespace NetConsts
-{
-    static int constexpr s_MaxPlayers = 8;
-    static int constexpr s_Port = 3148;
-}
